+++ src/cell/image.h
@@ -30,20 +30,21 @@
  *  POSSIBILITY OF SUCH DAMAGE.
 **/
 
-#ifndef CELL__IMAGE_H__INCLUDED
-#define CELL__IMAGE_H__INCLUDED
-
-#include "path.h"
+#ifndef SPHERE__IMAGE_H__INCLUDED
+#define SPHERE__IMAGE_H__INCLUDED
 
 #include <stdint.h>
 
+#include "fs.h"
+
 typedef struct image image_t;
 
-image_t*        image_open         (const path_t* path);
-void            image_close        (image_t* image);
-const uint32_t* image_get_pixelbuf (const image_t* image);
-const ptrdiff_t image_get_pitch    (const image_t* image);
-int             image_get_width    (const image_t* image);
-int             image_get_height   (const image_t* image);
+image_t*  image_load   (const fs_t* fs, const char* pathname);
+void      image_free   (image_t* image);
+int       image_height (const image_t* image);
+uint32_t* image_bitmap (const image_t* image, size_t *out_size);
+int       image_width  (const image_t* image);
+bool      image_save   (const image_t* image, const fs_t* fs, const char* filename);
+image_t*  image_slice  (const image_t* image, int x, int y, int width, int height);
 
-#endif // CELL__IMAGE_H__INCLUDED
+#endif // SPHERE__IMAGE_H__INCLUDED